Updates to keep up-to-date with dependencies

+++ examples/extensions/perturbationkernels/multivariate_normal_kernel.py
@@ -30,7 +30,7 @@ class MultivariateNormalKernel(PerturbationKernel, ContinuousKernel):
                             range(len(accepted_parameters_manager.kernel_parameters_bds.value()[kernel_index]))]
         for i in range(len(accepted_parameters_manager.kernel_parameters_bds.value()[kernel_index])):
             if isinstance(accepted_parameters_manager.kernel_parameters_bds.value()[kernel_index][i][0],
-                          (np.float, np.float32, np.float64, np.int, np.int32, np.int64)):
+                          (float, np.float32, np.float64, int, np.int32, np.int64)):
                 continuous_model[i] = accepted_parameters_manager.kernel_parameters_bds.value()[kernel_index][i]
             else:
                 continuous_model[i] = np.concatenate(
@@ -69,7 +69,7 @@ class MultivariateNormalKernel(PerturbationKernel, ContinuousKernel):
         continuous_model_values = accepted_parameters_manager.kernel_parameters_bds.value()[kernel_index]
 
         if isinstance(continuous_model_values[row_index][0],
-                      (np.float, np.float32, np.float64, np.int, np.int32, np.int64)):
+                      (float, np.float32, np.float64, int, np.int32, np.int64)):
             # Perturb
             cov = np.array(accepted_parameters_manager.accepted_cov_mats_bds.value()[kernel_index]).astype(float)
             continuous_model_values = np.array(continuous_model_values).astype(float)
@@ -114,7 +114,7 @@ class MultivariateNormalKernel(PerturbationKernel, ContinuousKernel):
             The pdf evaluated at point x.
         """
 
-        if isinstance(mean[0], (np.float, np.float32, np.float64, np.int, np.int32, np.int64)):
+        if isinstance(mean[0], (float, np.float32, np.float64, int, np.int32, np.int64)):
             mean = np.array(mean).astype(float)
             cov = np.array(accepted_parameters_manager.accepted_cov_mats_bds.value()[kernel_index]).astype(float)
             return multivariate_normal(mean, cov, allow_singular=True).pdf(np.array(x).astype(float))
